Create 'clients' Command
This is the command to get a list of active clients from Harvest.
Specifying --all returns all clients, except for those in the to be
excluded list. Quick details are in the README.

I disabled the Unit test because I haven't found a way to test an API
call that doesn't require 'real' credentials.