Implement #clients capability
Subject to change as development proceeds.

This first revision of #clients deals with client ids to exclude, as
well as --all or not.

I also changed #authenciated_user to a more appropriate API call.

#authenticate! was changed last minute because it dawned on me that
with each call we'd be creating a new client. I'm not sure it that
would've been bad, it just seems like that.